netsync: Initialize netsync package.
Create doc.go, interface.go, and README for new package.
