+++ netsync/README.md
@@ -1,0 +1,25 @@
+netsync
+=======
+
+[![Build Status](http://img.shields.io/travis/btcsuite/btcd.svg)](https://travis-ci.org/btcsuite/btcd)
+[![ISC License](http://img.shields.io/badge/license-ISC-blue.svg)](http://copyfree.org)
+[![GoDoc](https://img.shields.io/badge/godoc-reference-blue.svg)](http://godoc.org/github.com/btcsuite/btcd/netsync)
+
+## Overview
+
+This package implements a concurrency safe block syncing protocol. The provided
+implementation of SyncManager communicates with connected peers to perform an
+initial block download, keep the chain in sync, and announce new blocks
+connected to the chain. Currently the sync manager selects a single sync peer
+that it downloads all blocks from until it is up to date with the longest chain
+the sync peer is aware of.
+
+## Installation and Updating
+
+```bash
+$ go get -u github.com/btcsuite/btcd/netsync
+```
+
+## License
+
+Package netsync is licensed under the [copyfree](http://copyfree.org) ISC License.
