+++ netsync/blocklogger.go
@@ -1,4 +1,8 @@
-package main
+// Copyright (c) 2015-2017 The btcsuite developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package netsync
 
 import (
 	"sync"
